cli changes, check for endpoint argument
